Check the number of arguments to variadic at
Currently, we don't do any check of the sort, which means that you can
call `at` via its variadic interface with any number of parameters,
which may not match the number of input arguments the backend expects.
This commit ensures via SFINAE that the number of arguments is the same
as the coordinate dimensionality.