Add C++17 compatible support for constraints
C++ constraints based on concepts are a C++20 feature, but they would be
tremendously powerful for covfie. Therefore, this commit adds a sort of
middleground where we can optionally use concepts without breaking C++17
support. This works by defining a new macro `CONSTRAINT(concept)`, which
evaluates to a concept constraint on C++20 compilers, and evaluates
simply to the kind `typename` on older (C++17) compilers.