Add flag to require support for C++20 concepts
With our newfangled support for concepts, we have gained a powerful tool
for diagnostics, but the problem is that we need to make sure that they
are actually used and that we're not just failing silently. The new
`COVFIE_REQUIRE_CXX20` allows users to specify that C++20 concepts are a
requirement, not just a suggestion.