Refactor description of abstract vectors
As it was before, vector description classes carried both the type, the
size, and the way the vector was stored, but the storage should actually
be separate to support backends where the storage is not variable, such
as a CUDA texture backend. This commit removes the storage variable from
the descriptors to make them more flexible.