Make backend constructors more explicit
The previous constructor logic was a complete mess, and it's honestly a
miracle that it held out as long as it did, because implicit conversions
were being made left and right. This commit changes all backend
constructors to be explicit, so we do not have any more implicit
conversions; this also allows us to get rid of some messy legacy
constructors. Sadly, this requires us to get rid of the copyability
constraint in the backend concept, because that requires implicit
convertability.