Simplify transformer configuration types
Previously, each transformer had to define its own configuration type,
but this kind of sucks; sometimes you simply want to use an external
type, which also makes it easier to share configurations between
backends of the same type but with different backends. In this commit,
we make it so that some layers simply use existing types (like
`std::monostate`).