Use `nd_size` for strided configuration
Previously this used its own input type, which was slightly awkward. It
also required the method to have its own `get_size` method, which was
not very pleasant to use. This commit changes the strided transformation
layer so that it doesn't use its own type, but rather the standardized
n-dimensional size type.