Add layout simulation code as extra
This is a little bit of code which simulates the reciprocal throughput
of a variety of index calculation schemes and plots them. Not a part of
the covfie code itself, but a nice extra that serves as motivation for
having different schemes, and also it produces a nice figure for our
paper.