Increase specificity of certain constructors
The constructors for some of our transformers, the layouts in
particular, was set up in such a way that they basically took any input
and the user had to hope that it was actually possible to copy from
there. This, of course, is not great, as it naturally invites errors.
This commit changes those constructors so they use more elaborate SFINAE
to make sure that all the participating components have the required
capabilities.