Increase robustness of compile-time Morton masks
It turns out GCC is a lot worse at determining what can be computed at
compile time and what cannot, so it was failing to produce efficient
code for the PDEP version of the morton layout. This commit resolves
this issue through a bit of metaprogramming. It doesn't look good but I
think it works.