Find CUDA dependencies in installed CMake configs
As it stands, an installed version of covfie with the CUDA plugin
available will correctly try to link against the CUDA runtime in a
dependent project, but it won't actually look for it. This means that,
if the order of operations happens to be such that covfie is loaded
before CUDA, a configuration error will be thrown. This commit ensures
that the installed CMake configuration looks for the CUDA runtime.