Fix bug in the CMake install interface
We previously relied on `CMAKE_INSTALL_INCLUDEDIR` being set, but this
is not guaranteed unless `GNUInstallDirs`. CMake, of course, was quietly
swallowing this code and producing invalid installations. This commit
ensures that `GNUInstallDirs` is always loaded before the install
interfaces are defined, and also uses the `CMAKE_INSTALL_INCLUDEDIR`
variable more consistently.