Rework platform selection CMake code
Currently, the selected platforms determine which header files are
installed, but this is not a great design; it makes the installation
depend very heavily on the configuration of the project which makes
little sense for a header-only library. This commit reworks this design,
installing all headers regardless of the configuration. The platform
variables now only impact the tests, the examples, and the benchmarks.